+++ planetmint/commands/planetmint.py
@@ -259,6 +259,18 @@ def run_init(args):
 
 
 @configure_planetmint
+def run_migrate_up(args):
+    validator = Validator()
+    schema.migrate_up(validator.models.connection)
+
+
+@configure_planetmint
+def run_migrate_down(args):
+    validator = Validator()
+    schema.migrate_up(validator.models.connection)
+
+
+@configure_planetmint
 def run_drop(args):
     """Drop the database"""
 
@@ -362,6 +374,10 @@ def create_parser():
     subparsers.add_parser("init", help="Init the database")
 
     subparsers.add_parser("drop", help="Drop the database")
+
+    subparsers.add_parser("migrate_up", help="Migrate up")
+
+    subparsers.add_parser("migrate_down", help="Migrate down")
 
     # parser for starting Planetmint
     start_parser = subparsers.add_parser("start", help="Start Planetmint")